+++ src/main/java/net/lopymine/pe/compat/ip/IPotionParticleThing.java
@@ -1,0 +1,9 @@
+package net.lopymine.pe.compat.ip;
+
+public interface IPotionParticleThing {
+
+	boolean particleEffects$get();
+
+	void particleEffects$set(boolean bl);
+
+}