+++ src/main/java/net/lopymine/pe/mixin/ip/ParticleEngineAccessor.java
@@ -1,0 +1,17 @@
+package net.lopymine.pe.mixin.ip;
+
+//? if <=1.21.9 {
+
+/*import net.minecraft.client.particle.ParticleEngine;
+import net.minecraft.client.renderer.texture.TextureAtlas;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(ParticleEngine.class)
+public interface ParticleEngineAccessor {
+
+	@Accessor("textureAtlas")
+	TextureAtlas getTextureAtlas();
+
+}
+*///?}